moved patch file into osteo folder
