+++ pepper-apis/studybuilder-cli/src/main/java/org/broadinstitute/ddp/studybuilder/task/OsteoV2Updates.java
@@ -1,19 +1,7 @@
 package org.broadinstitute.ddp.studybuilder.task;
 
 import com.typesafe.config.Config;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoAboutChildV2;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoNewActivities;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoPrequalUpdate;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoAboutYouV2;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoAdultConsentFixes;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoConsentAddendumV2;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoConsentVersion2;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoDdp7601;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoFamilyHistoryReturnToDashboard;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoMRFv2;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoNewFamilyHistory;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoPDFv2;
-import org.broadinstitute.ddp.studybuilder.task.osteo.OsteoSomaticConsentAddendum;
+import org.broadinstitute.ddp.studybuilder.task.osteo.*;
 import org.jdbi.v3.core.Handle;
 
 import java.nio.file.Path;
